requirements.txt
langgraph>=1.0.7
langgraph-checkpoint-postgres>=3.0.0
langchain>=1.2.10
langchain-core>=1.2.17
langchain-groq>=1.1.2
langchain-anthropic>=1.3.4
langchain-huggingface>=1.2.1
langchain-qdrant>=1.1.0
qdrant-client>=1.14.0
langchain-text-splitters>=0.3.0
fastmcp>=3.1.0
langchain-mcp-adapters>=0.2.1
google-search-results>=2.4.0
fastapi[standard]>=0.115.0
uvicorn>=0.30.0
sse-starlette>=2.0.0
chainlit>=2.0.0
ragas>=0.4.3
deepeval>=2.0.0
langsmith>=0.6.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.27.0
pydantic>=2.9.0